+++ sources/libs/translation_module/include/translation_module/aliases.h
@@ -1,0 +1,9 @@
+#pragma once
+
+namespace trm {
+
+using JsonMap = std::map<std::string, Json::Value>;
+using StringMap = std::map<std::string, std::string>;
+using IntMap = std::map<std::string, int64_t>;
+
+}
