+++ sources/libs/translation_module/include/translation_module/cmd_translator.h
@@ -3,60 +3,12 @@
 #include <memory>
 #include <vector>
 #include <json/json.h>
-#include "translation_module.h"
+#include "aliases.h"
 #include "cmd_info.h"
+#include "translation_module.h"
+#include "fillers.h"
 
 namespace trm {
-
-using JsonMap = std::map<std::string, Json::Value>;
-using StringMap = std::map<std::string, std::string>;
-using IntMap = std::map<std::string, int64_t>;
-
-//
-// Fillers
-//
-class IFiller {
-public:
-    virtual void fill(const Json::Value &json) = 0;
-};
-
-class JsonFiller : public IFiller {
-public:
-    JsonFiller(const std::string &fieldName, JsonMap &jsonMap) :
-        fieldName(fieldName),
-        jsonMap(jsonMap)
-    {
-    }
-
-    void fill(const Json::Value &json) override
-    {
-        jsonMap[fieldName] = json;
-    }
-private:
-    std::string fieldName;
-    JsonMap &jsonMap;
-};
-
-class StringFiller : public IFiller {
-public:
-    StringFiller(const std::string &fieldName, StringMap &stringMap) :
-        fieldName(fieldName),
-        stringMap(stringMap)
-    {
-    }
-
-    void fill(const Json::Value &json) override
-    {
-        if (!json.isString()) {
-            throw std::runtime_error("Ожидался JSON с типом String.");
-        }
-        stringMap[fieldName] = json.asString();
-    }
-private:
-    std::string fieldName;
-    StringMap &stringMap;
-};
-
 
 class CmdBuilder {
 };
